guigui: call Layout again after handling input
Sometimes it is hard to determine which Layout and Update should be
used for a code to update widgets. To avoid confusions, let's call
Layout again and integrate such updating logic to Layout whenever
possible.

In the future, Update might be deprecated.

This mitigates the issue of glitches on opening context menus.

Updates #32